Timesheet_ai_agent/requirements.txt
streamlit
pandas
crewai
openai
pytesseract
Pillow
google-generativeai
python-dotenv
opencv-python
opencv-python-headless
matplotlib
